Address PR review on provenance layering and interactivity viewer.
