+++ cvs/lib/report/provenance.py
@@ -5,6 +5,8 @@ from __future__ import annotations
 import subprocess
 from typing import Any, List, Mapping, Tuple
 
+from cvs.core.image_display import format_image_display
+
 
 RunCardRow = Tuple[str, str, bool]
 
@@ -58,20 +60,6 @@ def format_git_ref(*, commit: str = "", branch: str = "", dirty: bool = False) -
     if dirty:
         ref = f"{ref} (dirty)"
     return ref
-
-
-def format_image_display(*, image_tag: str = "", image_digest: str = "", image_id: str = "") -> str:
-    digest = image_digest or image_id
-    if digest:
-        short = digest
-        if "@" in short:
-            short = short.split("@", 1)[1]
-        if short.startswith("sha256:") and len(short) > 19:
-            short = f"{short[:19]}\u2026"
-        if image_tag:
-            return f"{image_tag} @ {short}"
-        return short
-    return image_tag or "\u2014"
 
 
 def build_inference_report_provenance(
